requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torch-geometric>=2.4.0
dgl>=1.1.0
networkx>=2.8

# Data Science and Numerical Computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.1.0
scipy>=1.7.0

# Chemistry and Molecular Processing
rdkit>=2022.9.1
deepchem>=2.7.0

# Graph ML and OGB
ogb>=1.3.6

# Experiment Tracking and Logging
mlflow>=2.0.0
tensorboard>=2.10.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Configuration and Utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.64.0
click>=8.1.0

# Testing
pytest>=7.1.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0

# Code Quality
black>=22.6.0
isort>=5.10.0
flake8>=5.0.0
mypy>=0.971

# Documentation
sphinx>=5.1.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.18.0

# Development Tools
jupyterlab>=3.4.0
ipywidgets>=8.0.0
ipykernel>=6.15.0

# Optional: For accelerated training
# torch-audio>=0.12.0  # Uncomment if needed
# torch-vision>=0.13.0  # Uncomment if needed

# Optional: For distributed training
# torch-distributed>=0.1.0  # Uncomment if needed

# System Dependencies (install via conda/mamba if available)
# - cudatoolkit=11.8 (for GPU support)
# - openbabel (for molecular format conversion)
# - graphviz (for graph visualization)